requirements.txt
openai 
PyPDF2
streamlit
python-dotenv